Moved EventCreation classes and TransactionPool classes to event-creator-impl module
