+++ platform-sdk/consensus-event-creator-impl/build.gradle.kts
@@ -7,3 +7,14 @@ plugins {
 mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
 
 description = "Default Consensus Event Creator Implementation"
+
+testModuleInfo {
+    requires("com.hedera.node.hapi")
+    requires("com.swirlds.base.test.fixtures")
+    requires("com.swirlds.common.test.fixtures")
+    requires("com.swirlds.config.extensions.test.fixtures")
+    requires("org.hiero.base.utility.test.fixtures")
+    requires("org.junit.jupiter.api")
+    requires("org.junit.jupiter.params")
+    requires("org.mockito")
+}
