+++ platform-sdk/consensus-event-creator-impl/src/main/java/org/hiero/consensus/event/creator/impl/pool/TransactionPool.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.pool;
+package org.hiero.consensus.event.creator.impl.pool;
 
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import com.swirlds.component.framework.component.InputWireLabel;
