+++ platform-sdk/swirlds-platform-core/src/main/java/module-info.java
@@ -103,7 +103,6 @@ module com.swirlds.platform.core {
     opens com.swirlds.platform.cli to
             info.picocli;
 
-    exports com.swirlds.platform.components.transaction;
     exports com.swirlds.platform.event.preconsensus;
     exports com.swirlds.platform.gossip.sync.protocol;
     exports com.swirlds.platform.gossip;
@@ -121,7 +120,6 @@ module com.swirlds.platform.core {
     exports com.swirlds.platform.event.orphan;
     exports com.swirlds.platform.publisher;
     exports com.swirlds.platform.components.consensus;
-    exports com.swirlds.platform.pool;
     exports com.swirlds.platform.state.snapshot;
     exports com.swirlds.platform.state.service.schemas;
     exports com.swirlds.platform.state.service;
