+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/SyncManagerTest.java
@@ -19,11 +19,11 @@ import com.swirlds.platform.gossip.sync.SyncManagerImpl;
 import com.swirlds.platform.network.PeerInfo;
 import com.swirlds.platform.network.topology.NetworkTopology;
 import com.swirlds.platform.network.topology.StaticTopology;
-import com.swirlds.platform.pool.TransactionPoolNexus;
 import com.swirlds.platform.system.status.StatusActionSubmitter;
 import com.swirlds.platform.test.fixtures.addressbook.RandomRosterBuilder;
 import java.util.List;
 import java.util.Random;
+import org.hiero.consensus.event.creator.impl.pool.TransactionPoolNexus;
 import org.hiero.consensus.model.node.NodeId;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
